Reorganize a smidge
